+++ testsuites/tmtests/tm22/task1.c
@@ -1,6 +1,6 @@
 /*
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -114,9 +114,9 @@ rtems_task Low_task(
 )
 {
   rtems_id          id;
-  uint32_t    index;
-  uint32_t    count;
-  uint32_t    size;
+  uint32_t          index;
+  uint32_t          count;
+  size_t            size;
   rtems_status_code status;
 
   status = rtems_task_create(
